apps/backend/requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
aiofiles==23.2.1
asyncio
supabase==2.0.0
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3 